pyproject.toml
[project]
name = "pallas-bot"
version = "2.0.0"
description = "《明日方舟》帕拉斯 Bot"
authors = [{ name = "MistEO" }]
maintainers = [{ name = "mxcoras", email = "mxcoras@outlook.com" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "beanie>=1.30.0",
    "httpx>=0.27.0",
    "jieba>=0.42.1",
    "nb-cli>=1.4.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot2[fastapi]>=2.4.2",
    "pyncm-async>=0.1.3",
    "pypinyin>=0.54.0",
    "python-multipart>=0.0.20",
    "python-ulid>=3.0.0",
    "tenacity>=9.1.2",
]

[project.optional-dependencies]
perf = [
    "jieba-fast>=0.53",
]

[dependency-groups]
dev = [
    "pip-audit>=2.9.0",
    "ruff>=0.11.13",
]

[[tool.uv.index]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["src/plugins"]
